Moved permissionUtils class to permissions package
