+++ collect_app/src/main/java/org/odk/collect/android/activities/FillBlankFormActivity.java
@@ -50,7 +50,7 @@ import org.odk.collect.android.tasks.DiskSyncTask;
 import org.odk.collect.android.utilities.ApplicationConstants;
 import org.odk.collect.android.utilities.DialogUtils;
 import org.odk.collect.android.utilities.MultiClickGuard;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 import org.odk.collect.android.views.ObviousProgressBar;
 
 import javax.inject.Inject;
