+++ collect_app/src/main/java/org/odk/collect/android/activities/InstanceUploaderActivity.java
@@ -34,7 +34,7 @@ import org.odk.collect.android.utilities.ArrayUtils;
 import org.odk.collect.android.utilities.AuthDialogUtility;
 import org.odk.collect.android.utilities.DialogUtils;
 import org.odk.collect.android.utilities.InstanceUploaderUtils;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 
 import java.util.ArrayList;
 import java.util.Collections;
