+++ collect_app/src/main/java/org/odk/collect/android/activities/SplashScreenActivity.java
@@ -35,7 +35,7 @@ import org.odk.collect.android.preferences.GeneralSharedPreferences;
 import org.odk.collect.android.storage.StorageInitializer;
 import org.odk.collect.android.utilities.DialogUtils;
 import org.odk.collect.android.utilities.FileUtils;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
