+++ collect_app/src/main/java/org/odk/collect/android/configure/qr/QRCodeTabsActivity.java
@@ -21,7 +21,7 @@ import org.odk.collect.android.preferences.PreferencesProvider;
 import org.odk.collect.android.utilities.ActivityAvailability;
 import org.odk.collect.android.utilities.FileProvider;
 import org.odk.collect.android.utilities.MultiClickGuard;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 import org.odk.collect.async.Scheduler;
 
 import javax.inject.Inject;
