+++ collect_app/src/main/java/org/odk/collect/android/gdrive/GoogleDriveActivity.java
@@ -62,7 +62,7 @@ import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
 import org.odk.collect.android.utilities.DialogUtils;
 import org.odk.collect.android.utilities.FileUtils;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 
 import java.io.File;
 import java.io.IOException;
