+++ collect_app/src/main/java/org/odk/collect/android/logic/PropertyManager.java
@@ -26,7 +26,7 @@ import org.odk.collect.android.events.ReadPhoneStatePermissionRxEvent;
 import org.odk.collect.android.events.RxEventBus;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
 import org.odk.collect.android.utilities.DeviceDetailsProvider;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 
 import java.util.HashMap;
 import java.util.List;
