+++ collect_app/src/main/java/org/odk/collect/android/widgets/utilities/RecordingRequesterFactory.java
@@ -7,7 +7,7 @@ import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.formentry.FormEntryViewModel;
 import org.odk.collect.android.utilities.ActivityAvailability;
 import org.odk.collect.android.utilities.FormEntryPromptUtils;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 import org.odk.collect.android.utilities.QuestionMediaManager;
 import org.odk.collect.audiorecorder.recording.AudioRecorderViewModel;
