+++ collect_app/src/test/java/org/odk/collect/android/activities/support/AlwaysDenyStoragePermissionPermissionUtils.java
@@ -7,7 +7,7 @@ import androidx.annotation.NonNull;
 import org.odk.collect.android.listeners.PermissionListener;
 import org.odk.collect.android.permissions.PermissionsChecker;
 import org.odk.collect.android.storage.StorageStateProvider;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 
 public class AlwaysDenyStoragePermissionPermissionUtils extends PermissionUtils {
