+++ collect_app/src/test/java/org/odk/collect/android/fakes/FakePermissionUtils.java
@@ -7,7 +7,7 @@ import androidx.test.platform.app.InstrumentationRegistry;
 import org.odk.collect.android.listeners.PermissionListener;
 import org.odk.collect.android.permissions.PermissionsChecker;
 import org.odk.collect.android.storage.StorageStateProvider;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 
 /**
  * Mocked implementation of {@link PermissionUtils}.
