+++ collect_app/src/test/java/org/odk/collect/android/widgets/utilities/RecordingRequesterFactoryTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 import org.odk.collect.android.formentry.FormEntryViewModel;
 import org.odk.collect.android.support.MockFormEntryPromptBuilder;
 import org.odk.collect.android.utilities.ActivityAvailability;
-import org.odk.collect.android.utilities.PermissionUtils;
+import org.odk.collect.android.permissions.PermissionUtils;
 import org.odk.collect.android.utilities.QuestionMediaManager;
 import org.odk.collect.audiorecorder.recording.AudioRecorderViewModel;
